+++ circuits/examples/hybrid_mt.rs
@@ -338,6 +338,7 @@ impl Relation for HybridMtExample {
             secp256k1: false,
             bls12_381: false,
             base64: false,
+            nr_pow2range_cols: 1,
         }
     }