+++ circuits/examples/membership.rs
@@ -92,6 +92,7 @@ impl Relation for MembershipExample {
             secp256k1: false,
             bls12_381: false,
             base64: false,
+            nr_pow2range_cols: 1,
         }
     }