+++ circuits/src/field/decomposition/chip.rs
@@ -25,7 +25,7 @@ use num_traits::Zero;
 use super::{
     cpu_utils::{compute_optimal_limb_sizes, process_limb_sizes},
     instructions::CoreDecompositionInstructions,
-    pow2range::{Pow2RangeChip, Pow2RangeConfig, NB_POW2RANGE_COLS},
+    pow2range::{Pow2RangeChip, Pow2RangeConfig},
 };
 use crate::{
     field::{
@@ -71,7 +71,7 @@ impl P2RDecompositionConfig {
                 .zip(pow2range_config.val_cols.iter())
                 .all(|(n_col, p2r_col)| n_col == p2r_col),
             "DecompositionChip: Native and Pow2Range configs do not agree on the first {} columns",
-            NB_POW2RANGE_COLS
+            pow2range_config.val_cols.len()
         );
         Self {
             native_config: native_config.clone(),
@@ -132,7 +132,7 @@ impl<F: PrimeField> ComposableChip<F> for P2RDecompositionChip<F> {
         for bound in 0..=F::NUM_BITS {
             compute_optimal_limb_sizes(
                 &mut opt_limbs,
-                NB_POW2RANGE_COLS,
+                config.pow2range_config.val_cols.len(),
                 max_bit_len,
                 bound as i32,
             );
@@ -209,6 +209,8 @@ impl<F: PrimeField> P2RDecompositionChip<F> {
         x: Value<F>,
         limb_sizes: &[usize],
     ) -> Result<(AssignedNative<F>, Vec<AssignedNative<F>>), Error> {
+        let nr_pow2range_cols = self.pow2range_chip.config().val_cols.len();
+
         // assert limb_sizes structure is correct
 
         // 1. max limb length is not bigger than max_bit_length of chip
@@ -221,14 +223,14 @@ impl<F: PrimeField> P2RDecompositionChip<F> {
         // 2. the number of given limbs is multiple of NB_POW2RANGE_COLS
         #[cfg(not(test))]
         assert!(
-            limb_sizes.len() % NB_POW2RANGE_COLS == 0,
+            limb_sizes.len() % nr_pow2range_cols == 0,
             "Decomposition chip: number of limbs passed in decompose_core is not a multiple of NB_POW2RANGE_COLS",
         );
 
         // 3. each NB_POW2RANGE_COLS chunk is the same number and possibly some zeros
         #[cfg(not(test))]
         {
-            let limb_sizes_structure = limb_sizes.chunks(NB_POW2RANGE_COLS).all(|chunk| {
+            let limb_sizes_structure = limb_sizes.chunks(nr_pow2range_cols).all(|chunk| {
                 let mut v = chunk.to_vec();
                 v.sort();
                 v.dedup();
@@ -254,7 +256,7 @@ impl<F: PrimeField> P2RDecompositionChip<F> {
 
                 // compute the range_check tags for each column
                 let tags = limb_sizes
-                    .chunks(NB_POW2RANGE_COLS)
+                    .chunks(nr_pow2range_cols)
                     .map(|x| x[0])
                     .collect::<Vec<_>>();
 
@@ -279,7 +281,7 @@ impl<F: PrimeField> P2RDecompositionChip<F> {
                     terms.as_slice(),
                     F::ZERO,
                     &x,
-                    NB_POW2RANGE_COLS,
+                    nr_pow2range_cols,
                     &mut offset,
                 )?;
                 offset += 1;
@@ -337,13 +339,13 @@ impl<F: PrimeField> CoreDecompositionInstructions<F> for P2RDecompositionChip<F>
         if limb_size <= self.max_bit_len {
             // prepare the limb_size slice by filling with zeros to do parallel lookups
             let mut limb_sizes = vec![limb_size; number_of_limbs];
-            process_limb_sizes(NB_POW2RANGE_COLS, &mut limb_sizes);
+            process_limb_sizes(self.pow2range_chip.config().val_cols.len(), &mut limb_sizes);
             // prepare the limb sizes for last (possibly smaller limb). This is either empty
             // or contains exactly NB_POW2RANGE_COLS elements where the last
             // NB_POW2RANGE_COLS-1 are 0s
             if last_limb_size != 0 {
                 limb_sizes.push(last_limb_size);
-                process_limb_sizes(NB_POW2RANGE_COLS, &mut limb_sizes);
+                process_limb_sizes(self.pow2range_chip.config().val_cols.len(), &mut limb_sizes);
             }
 
             // we call the core function to retrieve the result
@@ -398,7 +400,7 @@ impl<F: PrimeField> CoreDecompositionInstructions<F> for P2RDecompositionChip<F>
                             terms.as_slice(),
                             F::ZERO,
                             &x.value().copied(),
-                            NB_POW2RANGE_COLS,
+                            self.pow2range_chip.config().val_cols.len(),
                             &mut offset,
                         )?;
 
@@ -437,7 +439,7 @@ impl<F: PrimeField> CoreDecompositionInstructions<F> for P2RDecompositionChip<F>
         // 2. process them by adding 0 terms in non-full rows
         optimal_limb_sizes
             .iter_mut()
-            .for_each(|row| process_limb_sizes(NB_POW2RANGE_COLS, row));
+            .for_each(|row| process_limb_sizes(self.pow2range_chip.config().val_cols.len(), row));
         let limb_sizes = optimal_limb_sizes.concat();
 
         // 3. use decompose_core to compute the result
@@ -459,12 +461,15 @@ impl<F: PrimeField> CoreDecompositionInstructions<F> for P2RDecompositionChip<F>
         );
 
         let mut assigned_values = Vec::with_capacity(values.len());
-        for chunk in values.chunks(NB_POW2RANGE_COLS) {
+        for chunk in values.chunks(self.pow2range_chip.config().val_cols.len()) {
             let assigned_chunk: Vec<_> = layouter.assign_region(
                 || "assign_many_small",
                 |mut region| {
                     let mut padded_chunk = chunk.to_vec();
-                    padded_chunk.resize(NB_POW2RANGE_COLS, Value::known(F::ZERO));
+                    padded_chunk.resize(
+                        self.pow2range_chip.config().val_cols.len(),
+                        Value::known(F::ZERO),
+                    );
 
                     self.pow2range_chip()
                         .assert_row_lower_than_2_pow_n(&mut region, bit_length)?;